Configure BDD specs, add details for running tests to README
[changelog]
added: information for running tests to README
added: tmpdir dependency for tests
added: basic BDD specs with critical-path coverage
added: better log output during runs
changed: set sane values for limits in example-config.yml
fixed: concurrency set to incorrect type in config.cr
fixed: ignore output binary from git

+++ src/config.cr
@@ -26,7 +26,7 @@ module S3t
   class LimitsConfig
     include YAML::Serializable
 
-    property concurrency : UInt8 = 10
+    property concurrency : Int32 = 10
     property count : UInt32
     property per_dir : UInt32
   end
